Modified: return shift schedule result if finish

+++ Data/Models/Requests/ShiftScheduleRequest.cs
@@ -2,6 +2,6 @@
 {
     public class ShiftScheduleRequest
     {
-        public long ShiftScheduleResultId { get; set; }
+        public int ShiftScheduleResultId { get; set; }
     }
 }
